컨트롤러와 관련 dto 구현
- 추가적으로 카카오 키워드 장소 검색 api 적용에 빠뜨린 코드 추가

+++ src/main/java/com/example/nearestconveniencestore/direction/entity/Direction.java
@@ -33,6 +33,6 @@ public class Direction extends BaseTimeEntity {
     private double targetLatitude;
     private double targetLongitude;
 
-    // 사용자 주소와 제설함 주소 사이의 거리
-    private double distance;
+    // 사용자 주소와 편의점 주소 사이의 거리
+    private int distance;
 }
